# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Directive OS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings:
    get:
      operationId: getListings
      tags: [listings]
      summary: Get all property listings for the agency
      responses:
        "200":
          description: List of listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
    post:
      operationId: createListing
      tags: [listings]
      summary: Create a new listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingBody"
      responses:
        "201":
          description: Created listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
  /listings/{id}:
    get:
      operationId: getListing
      tags: [listings]
      summary: Get a specific listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "404":
          description: Not found
    patch:
      operationId: updateListing
      tags: [listings]
      summary: Update a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateListingBody"
      responses:
        "200":
          description: Updated listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
    delete:
      operationId: deleteListing
      tags: [listings]
      summary: Delete a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /leads:
    get:
      operationId: getLeads
      tags: [leads]
      summary: Get all leads for the agency
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [new, contacted, qualified, closed]
        - name: type
          in: query
          schema:
            type: string
            enum: [buyer, tenant, landlord, vendor]
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a new lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadBody"
      responses:
        "201":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a specific lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update lead status or details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadBody"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
...
  /staff:
    get:
      operationId: getStaff
      tags: [staff]
      summary: Get all staff members in the agency
      responses:
        "200":
          description: List of staff
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"
    post:
      operationId: inviteStaff
      tags: [staff]
      summary: Invite a new staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteStaffBody"
      responses:
        "201":
          description: Invited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
  /staff/{id}:
    delete:
      operationId: removeStaff
      tags: [staff]
      summary: Remove a staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
...
components:
  schemas:
...
    Listing:
      type: object
      properties:
        id:
          type: integer
        agencyId:
          type: integer
        address:
          type: string
        suburb:
          type: string
        state:
          type: string
        postcode:
          type: string
        price:
          type: ["string", "null"]
        listingType:
          type: string
          enum: [sale, rental]
        bedrooms:
          type: ["integer", "null"]
        bathrooms:
          type: ["integer", "null"]
        agentName:
          type: string
        agentMobile:
          type: string
        inspectionTimes:
          type: array
          items:
            type: string
        vaultreId:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, under_offer, sold, leased]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - agencyId
        - address
        - suburb
        - state
        - postcode
        - listingType
        - agentName
        - agentMobile
        - inspectionTimes
        - status
        - createdAt
    CreateListingBody:
      type: object
      properties:
        address:
          type: string
        suburb:
          type: string
        state:
          type: string
        postcode:
          type: string
        price:
          type: string
        listingType:
          type: string
          enum: [sale, rental]
        bedrooms:
          type: integer
        bathrooms:
          type: integer
        agentName:
          type: string
        agentMobile:
          type: string
        inspectionTimes:
          type: array
          items:
            type: string
      required:
        - address
        - suburb
        - state
        - postcode
        - listingType
        - agentName
        - agentMobile
    UpdateListingBody:
      type: object
      properties:
        price:
          type: string
        status:
          type: string
          enum: [active, under_offer, sold, leased]
        inspectionTimes:
          type: array
          items:
            type: string
        agentMobile:
          type: string
...
    Lead:
      type: object
      properties:
        id:
          type: integer
        agencyId:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        leadType:
          type: string
          enum: [buyer, tenant, landlord, vendor, unknown]
        status:
          type: string
          enum: [new, contacted, qualified, closed]
        listingId:
          type: ["integer", "null"]
        listingAddress:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        formRequested:
          type: boolean
        hotLead:
          type: boolean
        channel:
          type: string
          enum: [voice, chat, web]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - agencyId
        - name
        - leadType
        - status
        - formRequested
        - hotLead
        - channel
        - createdAt
    CreateLeadBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        leadType:
          type: string
          enum: [buyer, tenant, landlord, vendor, unknown]
        listingId:
          type: integer
        notes:
          type: string
        channel:
          type: string
          enum: [voice, chat, web]
      required:
        - name
        - leadType
        - channel
    UpdateLeadBody:
      type: object
      properties:
        status:
          type: string
          enum: [new, contacted, qualified, closed]
        notes:
          type: string
        hotLead:
          type: boolean
        formRequested:
          type: boolean
...
    StaffMember:
      type: object
      properties:
        id:
          type: integer
        agencyId:
          type: integer
        clerkUserId:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [principal, agent]
        status:
          type: string
          enum: [active, invited, inactive]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - agencyId
        - clerkUserId
        - name
        - email
        - role
        - status
        - createdAt
    InviteStaffBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [principal, agent]
      required:
        - name
        - email
        - role
...