# lib/api-zod/src/generated/types/createListingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Directive OS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateListingBodyListingType } from "./createListingBodyListingType";

export interface CreateListingBody {
  address: string;
  suburb: string;
  state: string;
  postcode: string;
  price?: string;
  listingType: CreateListingBodyListingType;
  bedrooms?: number;
  bathrooms?: number;
  agentName: string;
  agentMobile: string;
  inspectionTimes?: string[];
}